Introduce a set of classes with Wildcard interface to encapsulate string matching
This should offload optimizations and selection of strategy to the moment of
pattern creation instead of usage saving on cpu cycles.

Especially important case is detecting if a pattern is wildcard or not takes 2
linear passes through each string + another one to actually compare to candidate.
By moving the first 2 passes to construction time this effect will not propagate to
O(N^2/N^3) algorithms down the line. The algorithms should be probably replaced as
well but that's a fight for another day.