Fix LDAP authentication when using StartTLS
When authenticating users against an LDAP directory that demand StartTLS
connection, the security plugin does not issue a StartTLS command to
upgrade the connection, leading to authentication failures.

The code use the ldaptive library to communicate with LDAP, but does not
use the same API as the one shown in the official ldaptive
documentation.  When used against an LDAP directory that demand
StartTLS, the examples of ldaptive perform the StartTLS upgrade
transparently and the communication with the directory is successful.

Update the code to use the same ldaptive API as the one of the working
examples.  This unbreak LDAP authentification when StartTLS is required.

As a bonus, remove the passing of the now useless properties all around
the code.

Signed-off-by: Romain Tartière <romain@blogreen.org>