Remove standalone admin tools zip file (#1628)
* Remove standalone admin tools zip file

There is a copy of the admin tools that are produced on build but are
not consumed or distributed as part of this project.  Until we have
documentation and/or tooling that uses this we should remove it from the
project.

Signed-off-by: Peter Nied <petern@amazon.com>
(cherry picked from commit 9f453d2d80e30b492045887aa47ca9f349aec4ff)