Add missing settings to plugin allowed list
Created FilterEntries enum, as the filter keys/namespaced-keys are used
in more than once spot if we add a new filter key, we need a way to know
if the entries were updated in the otherside of the codebase where the
defaults are used.  In the settings configuration, if a new enum value
is added and not included a runtime exception will fire on startup,
ensuring new settings are always included.

This could be better where the enum defines the default values and how
the settings are extracted from json, but that would be a considerable
larger scope with the inclusion of generics.

Signed-off-by: Peter Nied <petern@amazon.com>