Adopt new changelog standard
Added new human readable changelog for 2.0.0-rc1 and including
unreleased commits.

The release process we've been using for release-notes has not been
producing useful results, there has been discussion in
https://github.com/opensearch-project/OpenSearch/issues/1868 to come up
with a practice.  This is an attempt to start following the suggestion
and come up with feedback for the OpenSearch organization.

Signed-off-by: Peter Nied <petern@amazon.com>