New test framework (#1967)
This is the first iteration of a new framework for integration tests. The main focus is on conciseness and self-contained test. The framework offers:

A programmatic way of defining the cluster setup
A programmatic way of defining users and roles
A programmatic way of defining indices
A REST client with methods tailored for testing

This is now all part of the test class, so you do not need to jump around in various configuration files to get an overview on the test setup. There are of course still a lot of features missing, like setting up test data. I propose to add these features in increments.

Signed-off-by: Nils Bandener <nils.bandener@eliatra.com>
Signed-off-by: Jochen Kressin <jkressin@kressin.info>
Signed-off-by: Lukasz Soszynski <lukasz.soszynski@eliatra.com>
Signed-off-by: Kacper Trochimiak <kacper.trochimiak@eliatra.com>
Signed-off-by: Nils Bandener <git@bndnr.de>
Signed-off-by: Peter Nied <petern@amazon.com>
Co-authored-by: Jochen Kressin <jkressin@kressin.info>
Co-authored-by: Lukasz Soszynski <lukasz.soszynski@eliatra.com>
Co-authored-by: Kacper Trochimiak <kacper.trochimiak@eliatra.com>
Co-authored-by: Peter Nied <petern@amazon.com>