Keep version upgrade test working past v2
This change locks the version update to a known good 2.1.0-snapshot so
no matter the future version numbers being used, this test scenario will
still function.

When the version update is applied, it expects the build system for the
future version to exist, otherwise `gradle properties` will fail to
load.  This was causing issues when v3 updates were expecting v4
versions of opensearch build tools.

Signed-off-by: Peter Nied <petern@amazon.com>