Remove upgrade tests that keeps breaking on env issues
The upgrade check tests were a great idea that ultimately exposes some
problems with build compabtility between versions.  While we could
upgrade the test to re-install gradle after the upgrade since we are
effectively 'down grading' this is a really complex way to confirm that
code we don't often mess with works.

Signed-off-by: Peter Nied <petern@amazon.com>