Support multitenancy for the anonymous user
The current version of the plugin supports multitenancy
for all the users but not for the anonymous user.
The header 'securitytenant' is ignored when this user
is authorized. Therefore, all the operations run
with it, use the default tenant - normally the 'Global'
tenant -.

This patch fixes this issue assigning the selected
tenant - defined by the 'securitytenant' header -
to the anonymous user.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>
Signed-off-by: Eva Millán <evamillan@bitergia.com>

Co-authored-by: Eva Millán <evamillan@bitergia.com>