Support multitenancy for the anonymous user (#2459)
The current version of the plugin supports multitenancy
for all the users but not for the anonymous user.
The header 'securitytenant' is ignored when this user
is authorized. Therefore, all the operations run
with it, use the default tenant - normally the 'Global'
tenant -.

This patch fixes this issue assigning the selected
tenant - defined by the 'securitytenant' header -
to the anonymous user.

Co-authored-by: Eva Millán <evamillan@bitergia.com>
(cherry picked from commit 08e21dd0e04c9b92c47131bc60a3de3e4043e22d)