Limit the use of extensions term in the codebase
Extensions are a big part of what we are building into OpenSearch and
Security.  While we need these features we should try to build them as
generically as possible, this checkstyle scan will help enforce this and
also allow for working around the term usage with comments as we do have
places where extensions need to be treated differently.

Signed-off-by: Peter Nied <petern@amazon.com>