Removed unneeded task dependencies; Fixed copyExtraTestResources task
We've decided to remove dependency between `citest` and new, generated, test
tasks. This is not needed for CI runs as they will be always started in
separate Github Jobs. This is not significant change - it will only reduce
complexity a bit. This also required to change order of tasks in
build.gradle file.

This commit also fixes issue with copyExtraTestResources task. It's paths were
hardcoded which was fine for default test task but did not work for every
other, newly created, tasks. We've also added dependencies between all new
tasks and copyExtraTestResources

Signed-off-by: Pawel Gudel <pawel.gudel@eliatra.com>