[Backport 2.x][Build Break] Disable gradlew build cache to ensure most up-to-date dependencies
Backport 5e8f12c from #3186.

Disable gradlew build cache to ensure most up-to-date dependencies

We suspect that we are seeing the snapshot build of opensearch be cached
in some circumstances and not cached in others that is creating a mix of
build failures that are hard to interpret.

Follow up issue to revisit and potentially renable this setting
https://github.com/opensearch-project/security/issues/3185

Signed-off-by: Peter Nied <petern@amazon.com>
(cherry picked from commit 5e8f12ce5afe95f2f510cddf2a5b2cf50c076a66)