[Enchancement] Removed dependency between "test" and "jacocoTestReport" (#2935)
As a part of discussions in #2861 there's been determined that CI
workflow in Github Actions is excluding a lot of unnecessary task.
@nibix created #2913 to address this problem. Most of the unneeded tasks
has been already removed in #2861. This PR removes last existing part.
@peternied's confirmed removal of dependency in his
[comment](https://github.com/opensearch-project/security/issues/2913#issuecomment-1611985104).

This change removes the dependency between "_jacocoTestReport_" and
"_test_" tasks. Specifically the dependency that forces to start
"_test_" task always when "_jacocoTestReport_" is called. This allows us
to always generate coverage report in the end of every task run in "CI"
workflow without having to exclude "_test_" task in every single one of
them. Unfortunately this will break functionality to create code
coverage by starting only "_jacocoTestReport_" task. It looks like this
was not used widely and was certainly not used in any of Github Actions
Workflows. This will not break creating coverage reports in the end of
"_test_" task.


Signed-off-by: Pawel Gudel <pawel.gudel@eliatra.com>