Remove jarhell check disablement
JarHell was disabled due to an issue with Kafka where there were
multiple copies of the same class packaged inside of a release and test
distribution jar.  This was resolved with https://github.com/apache/kafka/pull/12407

The Kafka change has been merged, released, and pulled into this
repository so we are removing the workaround to supress jar hell issue
detection.

- Resolves #1938

Signed-off-by: Peter Nied <petern@amazon.com>