Increase coverage target
In RestLayerPrivilegesEvaluator there was a number of unused parameters
removed those to trim down the test requirements.

For the test cases switched to assertThat for easy of reading, enabled
debug logging to exercise debug logging code, adjusted usages of mocks
for cleaner test flow in a couple of spaces.

Verified locally with the following command and then checked the jacoco
report under build/reports/jacoco/test/html/...

./gradlew jacocoTestReport test --tests \
  org.opensearch.security.privileges.RestLayerPrivilegesEvaluatorTest

Signed-off-by: Peter Nied <petern@amazon.com>