Add RenameFieldResponseProcessor FLS/DLS tests (#3562)
### Description
Adds FLS/DLS tests for RenameFieldResponseProcessor
* Category Enhancement
* Why these changes are required? Testing to ensure the
RenameFieldResponseProcessor complies with FLS/DLS security
* What is the old behavior before changes and new behavior after
changes? No difference

### Issues Resolved
#2890

Is this a backport? If so, please add backport PR # and/or commits #:
No, but it does need to be backported.

### Testing
N/A

### Check List
- [x] New functionality includes testing
- [ ] New functionality has been documented
- [x] Commits are signed per the DCO using --signoff

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

---------

Signed-off-by: Sean Li <lnse@amazon.com>