Integrates OpenSAML 4.3 with SAML authenticator (#3651)
In order to exclude `permission
org.opensearch.secure_sm.ThreadPermission "modifyArbitraryThread";` we
must use the dedicated `cleaners` thread factory in `OpenSearch`. Since
`OpenSAML` packages are sealed it it impossible to replace
`CleanerSupport` class with our own solution. Instead we have to change
configuration of how such objects should be parsed. There are only 2
classes in the library which use `CleanerSupport`:
 - `X509CertificateImpl`
 - `X509CRLImpl` 

This fix uses the same solution as in `OpenSAML` and replaces
`CleanerSupport` with our custom implementation `CleanerFactory`.


Signed-off-by: Andrey Pleskach <ples@aiven.io>