Switch UserInjectorPlugin to use reliable per-instance settings (#3760)
### Description

Fix flaky tests in TransportUserInjectorIntegTest by removing reliance
on a static variable that makes the test unpredictable. This instead
adds a setting in the UserInjectorPlugin and relies on each node that
has the plugin installed to supply this setting when instantiating the
node.

* Category (Enhancement, New feature, Bug fix, Test fix, Refactoring,
Maintenance, Documentation)

Test Fix

### Issues Resolved

- Fixes https://github.com/opensearch-project/security/issues/3729

### Check List
- [ ] New functionality includes testing
- [ ] New functionality has been documented
- [ ] Commits are signed per the DCO using --signoff

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

Signed-off-by: Craig Perkins <cwperx@amazon.com>
Signed-off-by: Prabhas Kurapati <prabhask@berkeley.edu>