Refactor RestApiComplianceAuditlogTest to use TestAuditLogImpl.doThenWaitForMessages (#3733)
### Description

Refactors RestApiComplianceAuditlogTest to make this test suite more
stable and re-enable an ignored test.

The primary problem in this test suite is that messages that are auto
populated on startup get polluted with other audit log messages
generated when a user calls REST APIs. To make this more stable, a sleep
was added after `setup` to ignore the auto init messages and only
isolate messages generated on the API calls in each test.

`testAutoInit` is also updated to make it stable. There are 4 automatic
messages generated on cluster startup that have to do with the seeding
of the security index and audit log index.

* Category (Enhancement, New feature, Bug fix, Test fix, Refactoring,
Maintenance, Documentation)

Test Fix

### Issues Resolved

https://github.com/opensearch-project/security/issues/3730

### Check List
- [ ] New functionality includes testing
- [ ] New functionality has been documented
- [ ] Commits are signed per the DCO using --signoff

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

---------

Signed-off-by: Craig Perkins <cwperx@amazon.com>