[Maintenance-2930] Review use of JSON Flattener (#3674)
### Description

Implement JsonFlattener helper class as written in #2926 to deprecate
the use of the unnecessary JsonFlattener third party module.

* Category (Enhancement, New feature, Bug fix, Test fix, Refactoring,
Maintenance, Documentation)

Maintenance

* Why these changes are required?

The JsonFlattener module was being utilized in only one place for one
specific purpose, so these functions can be implemented as part of the
OpenSearch codebase instead of importing an unnecessary third party
module.

* What is the old behavior before changes and new behavior after
changes?

Hopefully nothing.

### Issues Resolved
- #2930

Is this a backport? If so, please add backport PR # and/or commits #

No

### Testing
Tests checked to make sure functions are not broken.

### Check List
- [x] New functionality includes testing
- [x] New functionality has been documented
- [x] Commits are signed per the DCO using --signoff

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

---------

Signed-off-by: Prabhas Kurapati <prabhask@berkeley.edu>