Resolve logback-classic to 1.2.13 (#3823)
### Description
Resolve logback-classic to 1.2.13 to resolve
https://github.com/advisories/GHSA-gm62-rw4g-vrc4

### Issues Resolved
Fix: #3821

Is this a backport? If so, please add backport PR # and/or commits #

### Testing
[Please provide details of testing done: unit testing, integration
testing and manual testing]

### Check List
- [ ] New functionality includes testing
- [ ] New functionality has been documented
- [ ] Commits are signed per the DCO using --signoff

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

Signed-off-by: Derek Ho <dxho@amazon.com>
Signed-off-by: Prabhas Kurapati <prabhask@berkeley.edu>