Improve messaging on how to set initial admin password (#3879)
### Description
When the demo configuration script fails when no initial admin password
is set, it does not provide helpful output telling the user how to set
the initial admin password. This PR provides some helpful output
indicating what variable the user should modify in order for the setup
to succeed.

### Issues Resolved
[List any issues this PR will resolve]

Is this a backport? If so, please add backport PR # and/or commits #

### Testing
Changed test asserting output 

### Check List
- [ ] New functionality includes testing
- [ ] New functionality has been documented
- [x] Commits are signed per the DCO using --signoff

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and
signing off your commits, please check
[here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).

---------

Signed-off-by: Derek Ho <dxho@amazon.com>